Moving auto parallel pass to xla/service
