+++ xla/service/BUILD
@@ -677,6 +677,19 @@ xla_cc_test(
 )
 
 cc_library(
+    name = "auto_parallel",
+    srcs = ["auto_parallel.cc"],
+    hdrs = ["auto_parallel.h"],
+    deps = [
+        "//xla/service:hlo_pass", 
+        "@com_google_absl//absl/container:flat_hash_set",
+        "@com_google_absl//absl/status:status",
+        "@com_google_absl//absl/strings:string_view",
+        "@tsl//tsl/platform:logging",
+    ],
+)
+
+cc_library(
     name = "sharding_propagation",
     srcs = [
         "sharding_propagation.cc",
