+++ xla/service/auto_parallel.cc
@@ -1,6 +1,6 @@
 // TODO: add copyright 
 
-#include "xla/service/cpu/auto_parallel.h"
+#include "xla/service/auto_parallel.h"
 
 #include "absl/container/flat_hash_set.h"
 #include "absl/status/status.h"
